# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ent.Bill:
    properties:
      Amount:
        description: Amount holds the value of the "Amount" field.
        type: string
      Date:
        description: Date holds the value of the "Date" field.
        type: string
      edges:
        $ref: '#/definitions/ent.BillEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BillQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.BillEdges:
    properties:
      officer:
        $ref: '#/definitions/ent.Financier'
        description: Officer holds the value of the officer edge.
        type: object
      paytype:
        $ref: '#/definitions/ent.Paytype'
        description: Paytype holds the value of the paytype edge.
        type: object
      treatment:
        $ref: '#/definitions/ent.Unpaybill'
        description: Treatment holds the value of the treatment edge.
        type: object
    type: object
  ent.Department:
    properties:
      department:
        description: Department holds the value of the "department" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DepartmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DepartmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DepartmentEdges:
    properties:
      department2doctorinfo:
        description: Department2doctorinfo holds the value of the department2doctorinfo
          edge.
        items:
          $ref: '#/definitions/ent.Doctorinfo'
        type: array
      historytaking:
        description: Historytaking holds the value of the historytaking edge.
        items:
          $ref: '#/definitions/ent.Historytaking'
        type: array
    type: object
  ent.Doctorinfo:
    properties:
      doctorname:
        description: Doctorname holds the value of the "doctorname" field.
        type: string
      doctorsurname:
        description: Doctorsurname holds the value of the "doctorsurname" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DoctorinfoEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DoctorinfoQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      licensenumber:
        description: Licensenumber holds the value of the "licensenumber" field.
        type: string
      telephonenumber:
        description: Telephonenumber holds the value of the "telephonenumber" field.
        type: string
    type: object
  ent.DoctorinfoEdges:
    properties:
      department:
        $ref: '#/definitions/ent.Department'
        description: Department holds the value of the department edge.
        type: object
      educationlevel:
        $ref: '#/definitions/ent.Educationlevel'
        description: Educationlevel holds the value of the educationlevel edge.
        type: object
      officeroom:
        $ref: '#/definitions/ent.Officeroom'
        description: Officeroom holds the value of the officeroom edge.
        type: object
      prename:
        $ref: '#/definitions/ent.Prename'
        description: Prename holds the value of the prename edge.
        type: object
      treatment:
        description: Treatment holds the value of the treatment edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
  ent.Educationlevel:
    properties:
      edges:
        $ref: '#/definitions/ent.EducationlevelEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EducationlevelQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      level:
        description: Level holds the value of the "level" field.
        type: string
    type: object
...
  ent.Financier:
    properties:
      edges:
        $ref: '#/definitions/ent.FinancierEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FinancierQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.FinancierEdges:
    properties:
      bills:
        description: Bills holds the value of the bills edge.
        items:
          $ref: '#/definitions/ent.Bill'
        type: array
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
        type: object
    type: object
  ent.Gender:
    properties:
      Genderstatus:
        description: Genderstatus holds the value of the "Genderstatus" field.
        type: string
      edges:
        $ref: '#/definitions/ent.GenderEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the GenderQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.GenderEdges:
    properties:
      patientrecord:
        description: Patientrecord holds the value of the patientrecord edge.
        items:
          $ref: '#/definitions/ent.Patientrecord'
        type: array
    type: object
  ent.Historytaking:
    properties:
      bp:
        description: Bp holds the value of the "bp" field.
        type: integer
      datetime:
        description: Datetime holds the value of the "datetime" field.
        type: string
      edges:
        $ref: '#/definitions/ent.HistorytakingEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the HistorytakingQuery when eager-loading is set.
        type: object
      hight:
        description: Hight holds the value of the "hight" field.
        type: number
      id:
        description: ID of the ent.
        type: integer
      oxygen:
        description: Oxygen holds the value of the "oxygen" field.
        type: string
      pulse:
        description: Pulse holds the value of the "pulse" field.
        type: integer
      respiration:
        description: Respiration holds the value of the "respiration" field.
        type: integer
      symptom:
        description: Symptom holds the value of the "symptom" field.
        type: string
      temp:
        description: Temp holds the value of the "temp" field.
        type: number
      weight:
        description: Weight holds the value of the "weight" field.
        type: number
    type: object
  ent.HistorytakingEdges:
    properties:
      department:
        $ref: '#/definitions/ent.Department'
        description: Department holds the value of the department edge.
        type: object
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      patientrecord:
        $ref: '#/definitions/ent.Patientrecord'
        description: Patientrecord holds the value of the patientrecord edge.
        type: object
      symptomseverity:
        $ref: '#/definitions/ent.Symptomseverity'
        description: Symptomseverity holds the value of the symptomseverity edge.
        type: object
    type: object
...
  ent.Medicalrecordstaff:
    properties:
      Name:
        description: Name holds the value of the "Name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.MedicalrecordstaffEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MedicalrecordstaffQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
...
  ent.Nurse:
    properties:
      edges:
        $ref: '#/definitions/ent.NurseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the NurseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      nursinglicense:
        description: Nursinglicense holds the value of the "nursinglicense" field.
        type: string
      position:
        description: Position holds the value of the "position" field.
        type: string
    type: object
...
  ent.Officeroom:
    properties:
      edges:
        $ref: '#/definitions/ent.OfficeroomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the OfficeroomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      roomnumber:
        description: Roomnumber holds the value of the "roomnumber" field.
        type: integer
    type: object
...
  ent.Patientrecord:
    properties:
      Age:
        description: Age holds the value of the "Age" field.
        type: integer
      Allergic:
        description: Allergic holds the value of the "Allergic" field.
        type: string
      Birthday:
        description: Birthday holds the value of the "Birthday" field.
        type: string
      Bloodtype:
        description: Bloodtype holds the value of the "Bloodtype" field.
        type: string
      Date:
        description: Date holds the value of the "Date" field.
        type: string
      Disease:
        description: Disease holds the value of the "Disease" field.
        type: string
      Email:
        description: Email holds the value of the "Email" field.
        type: string
      Home:
        description: Home holds the value of the "Home" field.
        type: string
      Idcardnumber:
        description: Idcardnumber holds the value of the "Idcardnumber" field.
        type: integer
      Name:
        description: Name holds the value of the "Name" field.
        type: string
      Phonenumber:
        description: Phonenumber holds the value of the "Phonenumber" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.PatientrecordEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PatientrecordQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PatientrecordEdges:
    properties:
      gender:
        $ref: '#/definitions/ent.Gender'
        description: Gender holds the value of the gender edge.
        type: object
      historytaking:
        description: Historytaking holds the value of the historytaking edge.
        items:
          $ref: '#/definitions/ent.Historytaking'
        type: array
      medicalrecordstaff:
        $ref: '#/definitions/ent.Medicalrecordstaff'
        description: Medicalrecordstaff holds the value of the medicalrecordstaff
          edge.
        type: object
      patientrecordPatientrights:
        description: PatientrecordPatientrights holds the value of the PatientrecordPatientrights
          edge.
        items:
          $ref: '#/definitions/ent.Patientrights'
        type: array
      prename:
        $ref: '#/definitions/ent.Prename'
        description: Prename holds the value of the prename edge.
        type: object
      treatment:
        description: Treatment holds the value of the treatment edge.
        items:
          $ref: '#/definitions/ent.Treatment'
        type: array
    type: object
  ent.Patientrights:
    properties:
      PermissionDate:
        description: PermissionDate holds the value of the "PermissionDate" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PatientrightsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PatientrightsQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      insurance_id:
        type: integer
      patientrightstype_id:
        type: integer
    type: object
...
  ent.Paytype:
    properties:
      edges:
        $ref: '#/definitions/ent.PaytypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PaytypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      paytype:
        description: Paytype holds the value of the "paytype" field.
        type: string
    type: object
...
  ent.Prename:
    properties:
      edges:
        $ref: '#/definitions/ent.PrenameEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PrenameQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      prefix:
        description: Prefix holds the value of the "prefix" field.
        type: string
    type: object
...
  ent.Symptomseverity:
    properties:
      edges:
        $ref: '#/definitions/ent.SymptomseverityEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SymptomseverityQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      symptomseverity:
        description: Symptomseverity holds the value of the "symptomseverity" field.
        type: string
    type: object
  ent.SymptomseverityEdges:
    properties:
      historytaking:
        description: Historytaking holds the value of the historytaking edge.
        items:
          $ref: '#/definitions/ent.Historytaking'
        type: array
    type: object
  ent.Treatment:
    properties:
      datetime:
        description: Datetime holds the value of the "datetime" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TreatmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TreatmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      treatment:
        description: Treatment holds the value of the "treatment" field.
        type: string
    type: object
...
  ent.User:
    properties:
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      financier:
        $ref: '#/definitions/ent.Financier'
        description: Financier holds the value of the financier edge.
        type: object
      historytaking:
        $ref: '#/definitions/ent.Nurse'
        description: Historytaking holds the value of the historytaking edge.
        type: object
      medicalrecordstaff:
        $ref: '#/definitions/ent.Medicalrecordstaff'
        description: Medicalrecordstaff holds the value of the medicalrecordstaff
          edge.
        type: object
      user2doctorinfo:
        $ref: '#/definitions/ent.Doctorinfo'
        description: User2doctorinfo holds the value of the user2doctorinfo edge.
        type: object
      userPatientrights:
        $ref: '#/definitions/ent.Patientrights'
        description: UserPatientrights holds the value of the UserPatientrights edge.
        type: object
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API
  version: "1.0"
paths:
  /Gender/{id}:
    get:
      description: get gender by ID
      operationId: get-gender
      parameters:
      - description: Gender ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Gender'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a gender entity by ID
...
  /department:
    get:
      description: list department entities
      operationId: list-department
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Department'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List department entities
    post:
      consumes:
      - application/json
      description: Create department
      operationId: create-department
      parameters:
      - description: Department entity
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/ent.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create department
  /department/{id}:
    delete:
      description: get department by ID
      operationId: delete-department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a department entity by ID
    get:
      description: get department by ID
      operationId: get-department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a department entity by ID
    put:
      consumes:
      - application/json
      description: update department by ID
      operationId: update-department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Department entity
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/ent.Department'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a department entity by ID
  /financier:
    get:
      description: list financier entities
      operationId: list-financier
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Financier'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List financier entities
  /financier/{id}:
    get:
      description: get financier by ID
      operationId: get-financier
      parameters:
      - description: Financier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Financier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a financier entity by ID
  /gender:
    get:
      description: list gender entities
      operationId: list-gender
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Gender'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List gender entities
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"